+++ src/RcsbGraphQL/GraphQLRequest.ts
@@ -1,5 +1,5 @@
 import {ApolloQueryResult} from "apollo-client";
-import ApolloClient from "apollo-boost";
+import ApolloClient, {PresetConfig} from "apollo-boost";
 import * as configBorregoGraphQL from "./ServerConfig/codegen.borrego.json";
 import * as configYosemiteGraphQL from "./ServerConfig/codegen.yosemite.json";
 import gql from 'graphql-tag';
@@ -9,20 +9,23 @@ export class GraphQLRequest {
 
     private readonly client: ApolloClient<unknown>;
 
-    constructor(api: "yosemite" | "borrego" | string) {
+    constructor(api: "yosemite" | "borrego" | string, config?: PresetConfig) {
         switch (api){
             case "yosemite":
                 this.client = new ApolloClient({
+                    ...config,
                     uri: configYosemiteGraphQL.schema
                 });
                 break;
             case "borrego":
                 this.client = new ApolloClient({
+                    ...config,
                     uri: configBorregoGraphQL.schema
                 });
                 break;
             default:
                 this.client = new ApolloClient({
+                    ...config,
                     uri: api
                 });
         }