+++ src/RcsbSearch/SearchRequest.ts
@@ -4,11 +4,14 @@ import * as serverSearch from "./ServerConfig/codegen.search.json";
 
 export class SearchRequest {
     private readonly uri: string;
-    constructor(uri?:string) {
+    private readonly fetch:(input:RequestInfo,init?:RequestInit)=>Promise<Response> = fetch;
+    constructor(uri?:string, externalFetch?:(input:RequestInfo,init?:RequestInit)=>Promise<Response>) {
         this.uri = uri ?? serverSearch.uri;
+        if(typeof externalFetch === "function")
+            this.fetch = externalFetch;
     }
-    public async request(query: SearchQuery): Promise<QueryResult>{
-        const response: Response = await fetch(this.uri, {
+    public async request(query: SearchQuery): Promise<QueryResult|null>{
+        const response: Response = await this.fetch(this.uri, {
             method: 'POST',
             headers: {
                 'Accept': 'application/json',